Rearrange for better readability
